@@ -1,22 +1,22 @@
 import 'package:flutter/material.dart';
-import 'package:votaciones2/pages/iniciarvotaciones.dart';
-import 'package:votaciones2/pages/personero.dart';
+import 'iniciarvotaciones.dart';
+import 'personero.dart';
 
 import 'archivo.dart';
 import "dart:io";
-
 
 final leeUsuario = TextEditingController();
 final leePassword = TextEditingController();
 final leenumtablet = TextEditingController();
 
-var usuario ="";
+var usuario = "";
 var password = "";
 var numtablet = "";
 
 class ingresar extends StatelessWidget {
-
-  const ingresar({Key? key, required this.usu, required this.pass, required this.curso}) : super(key: key);
+  const ingresar(
+      {Key? key, required this.usu, required this.pass, required this.curso})
+      : super(key: key);
   final String usu, pass, curso;
 
   @override
@@ -31,7 +31,9 @@ class ingresar extends StatelessWidget {
 }
 
 class cuerpo extends StatefulWidget {
-  const cuerpo({Key? key, required this.usu, required this.pass, required this.curso}) : super(key: key);
+  const cuerpo(
+      {Key? key, required this.usu, required this.pass, required this.curso})
+      : super(key: key);
   final String usu, pass, curso;
 
   @override
@@ -40,8 +42,9 @@ class cuerpo extends StatefulWidget {
   }
 }
 
-class cuerpoState extends State<cuerpo>{
-  cuerpoState({Key? key, required this.usu, required this.pass, required this.curso});
+class cuerpoState extends State<cuerpo> {
+  cuerpoState(
+      {Key? key, required this.usu, required this.pass, required this.curso});
   final String usu, pass, curso;
 
   @override
@@ -50,10 +53,9 @@ class cuerpoState extends State<cuerpo>{
     return Container(
       decoration: const BoxDecoration(
           image: DecorationImage(
-            image: AssetImage ('assets/images/escudoNC.jpg'),
+            image: AssetImage('assets/images/escudoNC.jpg'),
             fit: BoxFit.contain,
-          )
-      ),
+          )),
       child: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
@@ -62,64 +64,64 @@ class cuerpoState extends State<cuerpo>{
             campousuario(context),
             campocontrasena(context),
             campotablet(context),
-            const SizedBox(height: 10,),
+            const SizedBox(
+              height: 10,
+            ),
             botonentrar(context, usu, pass, curso)
           ],
         ),
-        ),
+      ),
     );
   }
 }
 
 Widget Sign_in(BuildContext context) {
-  return const Text("Iniciar Sesión",
-    style: TextStyle(color: Colors.black87, fontSize: 35.0, fontWeight: FontWeight.bold),);
-}
-
-Widget campousuario(BuildContext context){
-
-  var textController = TextEditingController();
-  return Container(
-      padding: const EdgeInsets.all(20),      //Darle bordes al TextField
-      child: TextField(
-
-          decoration: const InputDecoration(
-            enabledBorder: OutlineInputBorder(
-                borderSide: BorderSide(
-                  color: Colors.pink,
-                )
-            ),
-            labelText: "Nombre de usuario",
-            hintText: "Usuario",
-            prefixIcon: Icon(Icons.person),
-            fillColor: Colors.white,
-            filled: true,
-          ),
-        keyboardType: TextInputType.name,
-        controller: leeUsuario,
-      ),
+  return const Text(
+    "Iniciar Sesión",
+    style: TextStyle(
+        color: Colors.black87, fontSize: 35.0, fontWeight: FontWeight.bold),
   );
 }
 
-Widget campocontrasena(BuildContext context){
+Widget campousuario(BuildContext context) {
+  var textController = TextEditingController();
+  return Container(
+    padding: const EdgeInsets.all(20), //Darle bordes al TextField
+    child: TextField(
+      decoration: const InputDecoration(
+        enabledBorder: OutlineInputBorder(
+            borderSide: BorderSide(
+              color: Colors.pink,
+            )),
+        labelText: "Nombre de usuario",
+        hintText: "Usuario",
+        prefixIcon: Icon(Icons.person),
+        fillColor: Colors.white,
+        filled: true,
+      ),
+      keyboardType: TextInputType.name,
+      controller: leeUsuario,
+    ),
+  );
+}
 
+Widget campocontrasena(BuildContext context) {
   return Container(
       padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 5),
       child: TextField(
-          obscureText: true,
-          decoration: InputDecoration(
-            hintText: "Contraseña",
-            prefixIcon: Icon(Icons.lock),
-            fillColor: Colors.white,
-            filled: true,
-          ),
+        obscureText: true,
+        decoration: InputDecoration(
+          hintText: "Contraseña",
+          prefixIcon: Icon(Icons.lock),
+          fillColor: Colors.white,
+          filled: true,
+        ),
         keyboardType: TextInputType.name,
         controller: leePassword,
-      )
-  );
+      ));
 }
 
-Widget campotablet(BuildContext context){
+Widget campotablet(BuildContext context) {
   return Container(
       padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 5),
       child: TextField(
@@ -130,14 +132,16 @@ Widget campotablet(BuildContext context){
           filled: true,
         ),
         controller: leenumtablet,
-      )
-  );
+      ));
 }
 
-Widget botonentrar(BuildContext context, String usu, String pass, String curso) {
-  return FlatButton(
-      color: Colors.blueAccent,
-      padding: EdgeInsets.symmetric(horizontal: 30, vertical: 10),
+Widget botonentrar(
+    BuildContext context, String usu, String pass, String curso) {
+  return TextButton(
+      style: TextButton.styleFrom(
+        padding: EdgeInsets.symmetric(horizontal: 30, vertical: 10),
+        backgroundColor: Colors.blueAccent,
+      ),
       onPressed: () async {
         usuario = leeUsuario.text;
         password = leePassword.text;
@@ -149,11 +153,11 @@ Widget botonentrar(BuildContext context, String usu, String pass, String curso) 
           leenumtablet.text = '';
           Navigator.push(
             context,
-            MaterialPageRoute(builder: (context) => iniciarvotaciones(archivo: archivo)),
+            MaterialPageRoute(
+                builder: (context) => iniciarvotaciones(archivo: archivo)),
           );
         }
       },
-      child: Text(
-          "Ingresar", style: TextStyle(fontSize: 25, color: Colors.white))
-  );
+      child: Text("Ingresar",
+          style: TextStyle(fontSize: 25, color: Colors.white)));
 }
