@@ -1,9 +1,8 @@
-
 import 'package:flutter/material.dart';
-import 'package:votaciones2/pages/personero.dart';
+import 'personero.dart';
 import 'dart:io';
 
-import 'package:votaciones2/pages/sede.dart';
+import 'sede.dart';
 
 class iniciarvotaciones extends StatelessWidget {
   const iniciarvotaciones({Key? key, required this.archivo});
@@ -12,50 +11,53 @@ class iniciarvotaciones extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          title: const Text("INICAR VOTACIONES"),
-        ),
-        body: Center(
+      appBar: AppBar(
+        title: const Text("INICAR VOTACIONES"),
+      ),
+      body: Center(
           child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
-                RaisedButton(
-                  child: Text("INICIAR VOTACIÓN", style: TextStyle(fontSize: 25,)),
-                  onPressed: () {
+                ElevatedButton(
+                    child: Text("INICIAR VOTACIÓN",
+                        style: TextStyle(
+                          fontSize: 25,
+                        )),
+                    onPressed: () {
                       Navigator.push(
-                      context,
-                    MaterialPageRoute(builder: (context)=> personero(archivo: archivo)));
-                  }
-                ),
-              ]
-            )
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          showDialog(
-              barrierDismissible: false,                  //No quita el cuadro de Dialogo al darle clic por fuera de este
-              builder: (context){
-                return AlertDialog(
-                  title: Text("EL ARCHIVO DE VOTACION SE ENCUENTRA GRABADO EN..."),
-                  content: Text(archivo.toString()),
-                  actions: <Widget>[
-                    TextButton(
-                      onPressed: () {
-                        Navigator.push(
-                            context,
-                            MaterialPageRoute(builder: (context)=> const sede()));
-                      },
-                      child: Text("Aceptar"),
-                    )
-                  ],
-                );
-              },
-              context: context
-          );
-        },
-       tooltip: 'FINALIZAR VOTACIÓN',
-        child: const Icon(Icons.exit_to_app_rounded),
-    ),
+                          context,
+                          MaterialPageRoute(
+                              builder: (context) => personero(archivo: archivo)));
+                    }),
+              ])),
+          floatingActionButton: FloatingActionButton(
+            onPressed: () {
+              showDialog(
+                  barrierDismissible:
+                  false, //No quita el cuadro de Dialogo al darle clic por fuera de este
+                  builder: (context) {
+                    return AlertDialog(
+                      title:
+                      Text("EL ARCHIVO DE VOTACION SE ENCUENTRA GRABADO EN..."),
+                      content: Text(archivo.toString()),
+                      actions: <Widget>[
+                        TextButton(
+                          onPressed: () {
+                            Navigator.push(
+                                context,
+                                MaterialPageRoute(
+                                    builder: (context) => const sede()));
+                          },
+                          child: Text("Aceptar"),
+                        )
+                      ],
+                    );
+                  },
+                  context: context);
+            },
+            tooltip: 'FINALIZAR VOTACIÓN',
+            child: const Icon(Icons.exit_to_app_rounded),
+          ),
     );
   }
 }
