@@ -1,19 +1,12 @@
-//https://www.youtube.com/watch?v=RyIdsu4JUBo
-
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:votaciones2/pages/iniciarvotaciones.dart';
 
 import 'archivo.dart';
+import 'iniciarvotaciones.dart';
 import 'personero.dart';
 import 'dart:io';
 
 var fecha = DateTime.now();
-
-int _contadorcabildante1 = 0;
-int _contadorcabildante2 = 0;
-int _contadorcabildante3 = 0;
-int _contadorcabildante4 = 0;
 
 class cabildante extends StatelessWidget{
   cabildante({Key? key, required this.archivo, required this.person, required this.contra});
@@ -36,136 +29,50 @@ class cabildante extends StatelessWidget{
             crossAxisSpacing: 5,
             mainAxisSpacing: 5,
             children: <Widget> [
-              RaisedButton(
-                color: Colors.deepOrangeAccent,
-                child: Image.asset('assets/images/MariaValentinaGonzalez.jpg'),
-                onPressed: () {
-                  showDialog(
-                      barrierDismissible: false,                  //No quita el cuadro de Dialogo al darle clic por fuera de este
-                      builder: (context){
-                        return AlertDialog(
-                          title: Text("CODIGO DE VOTACIÓN"),
-                          content: Text(person + contra + cabil),
-                          actions: <Widget>[
-                            TextButton(
-                              onPressed: () {
-                                _contadorcabildante1++;
-                                cabil = "3";
-                                guardarvoto(archivo, person, contra, cabil);
-                                  Navigator.push(
-                                    context,
-                                    MaterialPageRoute(
-                                        builder: (context) => iniciarvotaciones(archivo: archivo)),
-                                  );
-                                },
-                                child: Text("Aceptar"),
-                            )
-                          ],
-                        );
-                      },
-                      context: context);
-                },
-              ),
-
-              RaisedButton(
-                color: Colors.pink[100],
-                child: Image.asset('assets/images/KarolValentinaDiaz.jpg'),
-                onPressed: () {
-                  showDialog(
-                      barrierDismissible: false,                  //No quita el cuadro de Dialogo al darle clic por fuera de este
-                      builder: (context){
-                        return AlertDialog(
-                          title: Text("CODIGO DE VOTACIÓN"),
-                          content: Text(person + contra + cabil),
-                          actions: <Widget>[
-                            TextButton(
-                              onPressed: () {
-                                cabil = "2";
-                                _contadorcabildante2++;
-                                guardarvoto(archivo, person, contra, cabil);
-                                Navigator.push(
-                                  context,
-                                  MaterialPageRoute(
-                                      builder: (context) => iniciarvotaciones(archivo: archivo)),
-                                );
-                              },
-                              child: Text("Aceptar"),
-                            )
-                          ],
-                        );
-                      },
-                      context: context);
-                },
-              ),
-
-              RaisedButton(
-                color: Colors.red,
-                child: Image.asset('assets/images/SharickMelissaMoreno.jpg'),
-                onPressed: () {
-                  showDialog(
-                      barrierDismissible: false,                  //No quita el cuadro de Dialogo al darle clic por fuera de este
-                      builder: (context){
-                        return AlertDialog(
-                          title: Text("CODIGO DE VOTACIÓN"),
-                          content: Text(person + contra + cabil),
-                          actions: <Widget>[
-                            TextButton(
-                              onPressed: () {
-                                cabil = "1";
-                                _contadorcabildante3++;
-                                guardarvoto(archivo, person, contra, cabil);
-                                Navigator.push(
-                                  context,
-                                  MaterialPageRoute(
-                                      builder: (context) => iniciarvotaciones(archivo: archivo)),
-                                );
-                              },
-                              child: Text("Aceptar"),
-                            )
-                          ],
-                        );
-                      },
-                      context: context);
-                },
-              ),
-
-              RaisedButton(
-                color: Colors.white,
-                child: Image.asset('assets/images/Votoenblanco.jpg'),
-                onPressed: () {
-                  showDialog(
-                      barrierDismissible: false,                  //No quita el cuadro de Dialogo al darle clic por fuera de este
-                      builder: (context){
-                        return AlertDialog(
-                          title: Text("CODIGO DE VOTACIÓN"),
-                          content: Text(person + contra + cabil),
-                          actions: <Widget>[
-                            TextButton(
-                              onPressed: () {
-                                cabil = "B";
-                                _contadorcabildante4++;
-                                guardarvoto(archivo, person, contra, cabil);
-                                Navigator.push(
-                                  context,
-                                  MaterialPageRoute(
-                                      builder: (context) => iniciarvotaciones(archivo: archivo)),
-                                );
-                              },
-                              child: Text("Aceptar"),
-                            )
-                          ],
-                        );
-                      },
-                      context: context
-                  );
-                },
-              ),
+              BotonCabildante(context, Colors.deepOrangeAccent, 'assets/images/MariaValentinaGonzalez.jpg', "3", archivo, person, contra),
+              BotonCabildante(context, Colors.pink[100], 'assets/images/KarolValentinaDiaz.jpg', "2", archivo, person, contra),
+              BotonCabildante(context, Colors.red, 'assets/images/SharickMelissaMoreno.jpg', "1", archivo, person, contra),
+              BotonCabildante(context, Colors.white, 'assets/images/Votoenblanco.jpg', "B", archivo, person, contra),
             ]
         ),
       ),
     );
   }
+
+  Widget BotonCabildante(BuildContext contexto, var color, String imagen, String numcabil, File archivo, String person, String contra) {
+
+    return ElevatedButton(
+      style: ElevatedButton.styleFrom(
+        backgroundColor: color,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30.0)),
+      ),
+      child: Image.asset(imagen),
+      onPressed: () {
+        showDialog(
+            barrierDismissible: false,                  //No quita el cuadro de Dialogo al darle clic por fuera de este
+            builder: (contexto){
+              return AlertDialog(
+                title: Text("CODIGO DE VOTACIÓN"),
+                content: Text(person + contra + cabil),
+                actions: <Widget>[
+                  TextButton(
+                    onPressed: () {
+                      cabil = numcabil;
+                      // llamar a la votacion del curso
+                      guardarvoto(archivo, person, contra, cabil); // falta agregar el nuevo parametro
+                      Navigator.push(
+                        contexto,
+                        MaterialPageRoute(
+                            builder: (context) => iniciarvotaciones(archivo: archivo)),
+                      );
+                    },
+                    child: Text("Aceptar"),
+                  )
+                ],
+              );
+            },
+            context: contexto);
+      },
+    );
+  }
 }
-
-
-
